pyproject.toml
[project]
name = "account-scanner"
version = "1.2.3"
description = "Multi-source account scanner: Reddit toxicity + Sherlock OSINT"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [{name = "Ven0m0"}]
keywords = ["osint", "reddit", "toxicity", "sherlock", "security"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.13",
  "Topic :: Security",
]

dependencies = [
  "httpx[http2]~=0.28.1",
  "orjson>=3.11.5,<4.0.0",
  "asyncpraw>=7.8.1,<8.0.0",
  "uvloop~=0.22.1; platform_system != 'Windows'",
  "aiofiles~=25.1",
  "discord.py>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff~=0.15.0",
  "mypy~=1.19.1",
  "pytest~=9.0.2",
  "pytest-asyncio~=1.3.0",
  "pip-audit>=2.10.0",  # Security vulnerability scanning
  "types-aiofiles>=24.0.0",  # Type stubs for mypy
]

[project.scripts]
account-scanner = "account_scanner:main"
scanner-bot = "discord_bot:main"

[build-system]
requires = ["setuptools>=78.1.1"]  # Security: CVE-2024-6345, PYSEC-2025-49
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["account_scanner", "discord_bot"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unreachable = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py313"
fix = true
exclude = [
  ".git",
  ".cache",
  ".var",
  ".venv",
  "venv",
  "node_modules",
  "__pycache__",
  "*.pyc",
]

[tool.ruff.lint]
select = [
  "E",     # pycodestyle errors
  "W",     # pycodestyle warnings
  "F",     # pyflakes
  "I",     # isort
  "N",     # pep8-naming
  "UP",    # pyupgrade
  "B",     # flake8-bugbear
  "C4",    # flake8-comprehensions
  "SIM",   # flake8-simplify
  "ASYNC", # flake8-async
  "TCH",   # flake8-type-checking
  "RUF",   # ruff-specific
]
ignore = [
  "E501",   # line too long (formatter handles)
  "SIM105", # contextlib.suppress (prefer explicit)
  "RUF005", # iterable unpacking (prefer simple concatenation for readability)
]

[tool.ruff.lint.isort]
known-first-party = ["account_scanner"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "."]
python_files = ["test_*.py"]
